+++ src-frontend/components/tools/ToolWidgets.tsx
@@ -22,3 +22,5 @@ export * from './SystemReminderWidget';
 export * from './SystemInitializedWidget';
 export * from './ThinkingWidget';
 export * from './SummaryWidget';
+export * from './ExitPlanModeWidget';
+export * from './ToolWithResultWidget';